accountListScreen.js
LWC for the First Screen - Displays a combobox with Industry options. Upon selection of an option, Accounts in that Industry will be displayed below along with the Phone number and Email columns. Along with these, the Accounts would contain a button "View Opportunities" which would take the user to the Second Screen.